Peanut Craft 3.0
This is the start of a new season:

Peanut Craft 3.0 is here!

And with it, comes tons of new, and awesome content we're all super excited to check out.

And if you're playing on the official server, that means a whole new world to check out!

In our testing, 2.8 worlds are compatible with 3.0. But there is no guarantee yours will be.

And there will be no support given for backwards compability. If you're not ready to move forward, then we suggest you stay on 2.8 for as long as you'd like.

But if you are ready to move forward, then welcome to the future!

And I bet you're curious about what new mods we've added? Well, take a look at this list:

Peanut Craft 2.8 to Peanut Craft 3.0

Added:
	- AppleCore
	- Baubley Heart Canisters
	- Clumps
	- Deep Mob Learning
	- Deep Mob Learning - Blood Magic Addon
	- Enchantment Descriptions
	- Engineers Workshop Reborn
	- Farming for Blockheads
	- Forgiving Void
	- InstantUnify
	- Matter Overdrive: Legacy Edition
	- Modular Powersuits
	- Nature Overhaul
	- Numina
	- Patchouli
	- Phosphor
	- Portality
	- Spice of Life: Carrot Edition
	- Stargate Network
	- Tech Reborn
	- Tinkering with Blood Magic
	- Too Many Efficiency Losses
	- Uppers

Removed:
	- Genetics Reborn (Too buggy/incomplete.)
	- Stargate Atlantiscraft (Replaced by Stargate Networks.)

* Generated using https://github.com/TheRandomLabs/ChangelogGenerator (1.13)

 We hope you're excited as we are. See you in the multi-verse!